\name{YPmodel-internal}
\alias{YPmodel-internal}
\alias{fun.adlgrk}
\alias{fun.classCheck}
\alias{fun.cumsum}
\alias{fun.errorMessage}
\alias{fun.flipud}
\alias{fun.generateParameters}
\alias{fun.hazardRationEstimation}
\alias{fun.hcvxitr1}
\alias{fun.internalParameters}
\alias{fun.martint}
\alias{fun.newBest}
\alias{fun.ntitr0}
\alias{fun.oldp2}
\alias{fun.Rjudge}
\alias{fun.survf}
\alias{fun.variance}
\alias{fun.YPmodelOriginal}
\alias{fun.zeroORone}
\alias{YPmodel.inputData}
\alias{YPmodel.setParameter}
\alias{YPmodel.setRandom}
\alias{summary.YPmodel.overall}
\title{Internal functions for Model of short-term and long-term hazard ratio for survival data  }
\description{
Internal functions for the YPmodel package.
}
\author{Junlong Sun and Song Yang}
\seealso{\code{\link{YPmodel}}}
\keyword{YPmodel}

